backend/pyproject.toml
[project]
name = "plannit-backend"
version = "0.1.0"
description = "Django backend for the Plannit Time Tracker application."
requires-python = ">=3.12,<3.13"
dependencies = [
  "Django==5.1.4",
  "dj-database-url==2.3.0",
  "django-cors-headers==4.6.0",
  "django-filter==24.3",
  "djangorestframework==3.15.2",
  "djangorestframework-simplejwt==5.3.1",
  "gunicorn==23.0.0",
  "psycopg[binary]==3.2.3",
  "python-dotenv==1.0.1",
  "whitenoise==6.8.2",
]

[dependency-groups]
dev = [
  "django-extensions==3.2.3",
  "ipython==8.30.0",
]

[tool.uv]
package = false
